requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioresponses==0.7.7
aiosignal==1.3.2
alpaca-trade-api==3.2.0
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.6.10
deprecation==2.1.0
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
lxml==5.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
textblob==0.18.0.post0
toml==0.10.2
tqdm==4.67.1
tzdata==2024.2
urllib3==1.26.20
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
yfinance==0.2.51
